requirements.txt
streamlit
openai
requests
yfinance
python-dotenv
openpyxl